\name{randtests-package}
\alias{randomtests-package}
\docType{package}
\title{
Testing randomness in R
}
\description{
Several non parametric tests for testing randomness in R.
}
\details{
\tabular{ll}{
Package: \tab randomtests\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-16\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Frederico Caeiro and Ayana Mateus

Maintainer: Frederico Caeiro <fac@fct.unl.pt>
}
