\name{randtests.aux}
\alias{randtests.aux}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliar function
}
\description{. 
}
\usage{
randtests.aux(x, method)
}
\arguments{
  \item{x}{data vector.}
  \item{method}{number of elements to choose. Default is \code{m=length(n)}.}
}
%\details{
%}
%\value{
%}
%\references{
%}
%\examples{
%}
%\keyword{  }
%\keyword{ randomness test }% __ONLY ONE__ keyword per line
\keyword{internal}