\name{sweetpotato}
\alias{sweetpotato}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sweet potato production
}
\description{
Sweetpotato Production, Yield per Acre and Acreage harvested in the United States, between 1868 and 1937. This data was already studied in Moore and Wallis (1943).
}
\format{
A list with 70 observations on 4 vectors: \code{year}, \code{production}, \code{yield} and \code{acreage}. 
}
\usage{
data(sweetpotato)
}
%- maybe also 'usage' for other objects documented here.
\source{
Agricultural Statistics 1939, p. 243. \url{http://archive.org/stream/agriculturalsat00unit#page/243/mode/1up}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ sweetpotato }
%\keyword{ randomness test }% __ONLY ONE__ keyword per line
