\name{randtests-package}
\alias{randtests-package}
\docType{package}
\title{
Testing randomness in R
}
\description{
The package \code{randtests} implements several nonparametric randomness tests of hypothesis. 
}
\details{
\tabular{ll}{
Package: \tab randomtests\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2024-04-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Randomness is a common assumption in many statistical methods. When such assumption is not fulfilled, we may draw wrong conclusions. Although in many datasets a simple graphical analysis is enough to check such assumption, in others a test of hypothesis is required.
}
\author{
Frederico Caeiro and Ayana Mateus

Maintainer: Frederico Caeiro <fac@fct.unl.pt>
}
\references{
Mateus A. and Caeiro F. (2014). An R implementation of several Randomness Tests. In T. E. Simos, Z. Kalogiratou and T. Monovasilis (eds.), \emph{AIP Conf. Proc.} \bold{1618}, 531--534.
}
